#14 - cleaned out html templates
