+++ licket-demo/src/main/java/org/licket/demo/view/ContactsList.java
@@ -7,10 +7,8 @@ import org.licket.core.view.list.AbstractLicketList;
 import org.licket.core.view.mount.MountedComponentLink;
 import org.licket.core.view.mount.params.MountingParamsAggregator;
 import org.licket.demo.model.Contact;
-import org.licket.demo.model.EmailAddress;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static org.licket.core.model.LicketComponentModel.ofString;
 import static org.licket.core.view.mount.params.MountingParamValueDecorator.fromParentModel;
 
 public class ContactsList extends AbstractLicketList<Contact> {
@@ -26,18 +24,6 @@ public class ContactsList extends AbstractLicketList<Contact> {
     protected void onInitializeContainer() {
         add(new LicketLabel("name"));
         add(new LicketLabel("description"));
-        add(new AbstractLicketList<EmailAddress>("email", ofString("emails")) {
-
-            @Override
-            protected LicketComponentModelReloader getModelReloader() {
-                return modelReloader;
-            }
-
-            @Override
-            protected void onInitializeContainer() {
-                add(new LicketLabel("value"));
-            }
-        });
         add(new MountedComponentLink<Contact>("view-contact", ViewContactPanel.class) {
             @Override
             protected void aggregateParams(MountingParamsAggregator paramsAggregator) {
