+++ licket-demo/src/main/java/org/licket/demo/view/ViewContactPanel.java
@@ -4,15 +4,18 @@ import org.licket.core.module.application.LicketComponentModelReloader;
 import org.licket.core.view.LicketLabel;
 import org.licket.core.view.container.AbstractLicketMultiContainer;
 import org.licket.core.view.hippo.vue.annotation.LicketMountPoint;
+import org.licket.core.view.list.AbstractLicketList;
 import org.licket.core.view.mount.MountedComponentLink;
 import org.licket.core.view.mount.params.MountingParams;
 import org.licket.demo.model.Contact;
+import org.licket.demo.model.EmailAddress;
 import org.licket.demo.service.ContactsService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Optional;
 
 import static org.licket.core.model.LicketComponentModel.ofModelObject;
+import static org.licket.core.model.LicketComponentModel.ofString;
 import static org.licket.core.view.LicketComponentView.fromComponentClass;
 
 /**
@@ -34,6 +37,21 @@ public class ViewContactPanel extends AbstractLicketMultiContainer<Contact> {
   @Override
   protected void onInitializeContainer() {
     add(new LicketLabel("name"));
+    add(new LicketLabel("description"));
+    add(new AbstractLicketList<EmailAddress>("email", ofString("emails")) {
+
+            @Override
+            protected LicketComponentModelReloader getModelReloader() {
+                return modelReloader;
+            }
+
+            @Override
+            protected void onInitializeContainer() {
+                add(new LicketLabel("value"));
+            }
+        });
+
+
     add(new MountedComponentLink("rootLink", ContactsAppRoot.class));
   }
