+++ licket-demo/src/main/resources/org/licket/demo/view/ViewContactPanel.html
@@ -1,7 +1,18 @@
-<div xmlns="http://www.w3.org/1999/xhtml" xmlns:lick="http://www.w3.org/1999/xhtml">
-    Here will be contact details
+<div class="ui container" xmlns="http://www.w3.org/1999/xhtml" xmlns:lick="http://www.w3.org/1999/xhtml">
 
-    <a href="#" lick:id="rootLink">go back to contacts list</a>
+    <div class="ui very padded segment">
+        <h2 class="ui header" lick:id="name">John Doe</h2>
+        <p lick:id="description"></p>
 
-    <div lick:id="name">???</div>
+        <div class="ui relaxed horizontal list">
+            <div class="item" lick:id="email">
+                <a href="#" lick:id="value">email@domain.com</a>
+            </div>
+        </div>
+
+        <button class="ui right floated labeled icon button" lick:id="rootLink">
+            <i class="arrow left"></i>
+            Back
+        </button>
+    </div>
 </div>
